+++ internal/habits/habits.go
@@ -6,13 +6,14 @@ import (
 
 	"github.com/utkuufuk/entrello/internal/config"
 	"github.com/utkuufuk/entrello/internal/trello"
+	"google.golang.org/api/sheets/v4"
 )
 
 type source struct {
 	spreadsheetId   string
 	credentialsFile string
 	tokenFile       string
-	service         service
+	service         *sheets.SpreadsheetsValuesService
 }
 
 func GetSource(cfg config.Habits) source {
@@ -20,11 +21,21 @@ func GetSource(cfg config.Habits) source {
 }
 
 func (s source) FetchNewCards(ctx context.Context, cfg config.SourceConfig) ([]trello.Card, error) {
-	err := s.createService(ctx)
+	err := s.initializeService(ctx)
 	if err != nil {
-		return nil, fmt.Errorf("could not create spreadsheet service: %w", err)
+		return nil, fmt.Errorf("could not initialize google spreadsheet service: %w", err)
 	}
+
 	data, _ := s.readCells("Jun 2020!B1:M1")
 	fmt.Println(data)
 	return []trello.Card{}, nil
+}
+
+// readCells reads a range of cell values with the given range
+func (s source) readCells(rangeName string) ([][]interface{}, error) {
+	resp, err := s.service.Get(s.spreadsheetId, rangeName).Do()
+	if err != nil {
+		return nil, fmt.Errorf("could not read cells: %w", err)
+	}
+	return resp.Values, nil
 }
